requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
pandas
python-dotenv
aiofiles